Update PitcherTeamApp.java
Added new labels and textboxes for each statistic. Validation for each textbox and each data field needs to be added to the report and formatted. 

As a result of adding more labels and textboxes, the gui needed to be wider. I made it center on the screen as it shows up now for it.